requirements.txt
crewai==0.5.0
python-dotenv>=1.0.1
langchain-google-genai==0.0.6
google-generativeai==0.3.2

